+++ Utility/Error.h
@@ -43,6 +43,7 @@ constexpr auto FATAL_BTP_SORT_WRITE_FAILED = RC_FATAL_BASE + 25;
 constexpr auto FATAL_SEEK_FAILED = RC_FATAL_BASE + 26;
 constexpr auto FATAL_READ_FAILED = RC_FATAL_BASE + 27;
 constexpr auto FATAL_FI_FLUSH_FAILED = RC_FATAL_BASE + 28;
+constexpr auto FATAL_TS_UNBALANCED_TREE = RC_FATAL_BASE + 29;
 
 /* Utility Implementation errors */
 constexpr auto  UTIL_RC_Success = RC_SUCCESS;